images/alpine: Fix edge build
The grub apk trigger fails when it can not find the rootfs device, and
the rootfs device is not found when executed in a container. Previously
apk (version 2.x) has exited with success, even if trigger fails, but
with apk certion 3.x, it now exits with failure, which caused the image
of alpine edge fail after apk was upgraded to 3.

Fix this by disable the grub apk trigger before the grub package is
installed, and rely on the distrobuilder's post-packages hook to
generate the grub config.

Re-enable it after again so future upgrades of grub in VM
will regenerate the config as needed.

Signed-off-by: Natanael Copa <ncopa@alpinelinux.org>
(cherry picked from commit 3bf8a4dff22caabe00b1eb9878a76e603fa39ec6)
Signed-off-by: Din Music <din.music@canonical.com>
License: Apache-2.0