github: do SFTP connectivity test in `actions/image-setup`
A `tcpdump` capture on the PS6 VM showed that failed upload were accompanied
with exactly 0 traffic from GH reaching the PS6 VM. Since this traffic is
subject to IP allow-lists, my current theory is that IS' list does not include
an up to date list of GH possible IPv4 source addresses.

As such, it's best to test the connectivity before doing anything "expensive"
with the GH runner. Also, when the connectivity will fail again, I'll be armed
with the public IPv4 that was used for the upload test and this will let me
correlate it with the allow-list maintained by IS.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>