github: move to a weekly build/publication cadence for `busybox`
Busybox is released as a single tarball once in a blue moon and there are not
package to update. The point of rebuilding/publishing new images is mostly to
pick up new images quickly but that is such a rare event that we can do a
manual build when that happens.

We could in theory stick to a single build and publish but we'll stick to a
weekly cadence to avoid the images to be considered stale and expunged from
`images:`.

This should reduce the pressure on the apparently overloaded download mirror.

Consumers of this image will also see less (pointless) refreshes which should
especially help if they use a storage driver that supports CoW.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>