tests/network-ovn: Host route for LB external IPs
Logical Routers in OVN 22.03 do not support answering ARP requests
for Load Balancer IP addresses that are not part of the subnet on one
of their Logical Router Ports[0].
This makes some of the tests in the `forward` suite flaky, because
these tests set up Load Balancers with external IP addresses that
don't belong to any of the router subnets. As such, they rely on
occasional left-over ARP entry for the LB IP to pass.

There's already an established precedent in the network-ovn test
suites to add a host route for the Load Balancer IP via the IP
address of the Logical Router Port.
This change adds the host routes in test suites where it was
missing.

[0] https://github.com/ovn-org/ovn/issues/124

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>